add breaking condoms interface. The clothingtype physicalattribute is a float fraction that a condom doesn't rip during penetration. If ripped, a condom receives the itemEffectType BROKEN_CONDOM that affects ignoreCondom() in the SexActioniInterface.

+++ src/com/lilithsthrone/controller/eventListeners/InventoryTooltipEventListener.java
@@ -867,7 +867,7 @@ public class InventoryTooltipEventListener implements EventListener {
 		// Attribute modifiers:
 		tooltipSB.append("<div class='container-full-width'>"
 				+ "<div class='container-half-width titular' style='width:calc(66.6% - 16px);'>");
-		int res = absClothing.getClothingType().getPhysicalResistance();
+		float res = absClothing.getClothingType().getPhysicalResistance();
 		tooltipSB.append(
 				(res>0
 					?"[style.boldGood(+"+absClothing.getClothingType().getPhysicalResistance()+")]"
